scenes

+++ renderer/sprite_object.h
@@ -9,7 +9,7 @@
 static const BString SPRITE_OBJECT_TYPE = BSTRING_CONST_LENGTH("Sprite", 6);
 void register_sprite_object();
 
-LS_EXPORT uint32 sprite_get_object_type();
+LS_EXPORT uint64 sprite_get_object_type();
 
 LS_EXPORT bool object_is_sprite(Object *object);
 LS_EXPORT Sprite *object_to_sprite(Object *object);
